refactor: split r2g/g2r into ir/emit layers and extract Go templates (task 2.3)
- Move the embedded C/Go template strings from rust2go-common's common.rs
  into standalone tmpl/ files loaded via include_str! (byte-exact, so the
  generated gen.go output is unchanged).
- Split r2g.rs into r2g/{mod,emit_go,emit_rust}.rs and g2r.rs into
  g2r/{mod,emit_c,emit_go,emit_rust}.rs: mod.rs keeps the IR (repr
  structs, syn parsing, accessors) while emitters are layered by target
  language. Public API paths and behavior are unchanged.
