+++ rust2go-common/src/tmpl/internal_drop.h.tmpl
@@ -1,0 +1,2 @@
+
+const void c_rust2go_internal_drop(void*);
