+++ rust2go-common/src/tmpl/structs_go118.go.tmpl
@@ -1,0 +1,25 @@
+
+        // An alternative impl of unsafe.String for go1.18
+        func unsafeString(ptr *byte, length int) string {
+            sliceHeader := &reflect.SliceHeader{
+                Data: uintptr(unsafe.Pointer(ptr)),
+                Len:  length,
+                Cap:  length,
+            }
+            return *(*string)(unsafe.Pointer(sliceHeader))
+        }
+
+        // An alternative impl of unsafe.StringData for go1.18
+        func unsafeStringData(s string) *byte {
+            return (*byte)(unsafe.Pointer((*reflect.StringHeader)(unsafe.Pointer(&s)).Data))
+        }
+        func newString(s_ref C.StringRef) string {
+            return unsafeString((*byte)(unsafe.Pointer(s_ref.ptr)), int(s_ref.len))
+        }
+        func refString(s *string, _ *[]byte) C.StringRef {
+            return C.StringRef{
+                ptr: (*C.uint8_t)(unsafeStringData(*s)),
+                len: C.uintptr_t(len(*s)),
+            }
+        }
+
